Day 4, part 2 - complete
The translation back to the standard coordinates to reverse the
diagonalising skew is a bit ugly.
I find the nesting of calls to map etc. a bit confusing, I can see why
the pipe operator is a good idea in functional languages, and it's a
pity erlang doesn't have it.
This code wasn't quick to develop, and it isn't particularly attractive,
so I'm not sure what objective it meets, other than me learning a bit!